Restructure chess project into modular workspace
