+++ projects/chess/tests/resources/empty_board.txt
