+++ projects/chess/tests/resources/inverted_moves.txt
