+++ projects/chess/tests/resources/inverted_starting_board.txt
