+++ projects/chess/tests/resources/non_standard_board.txt
