+++ projects/chess/tests/resources/pawn_knight_moves.txt
