+++ projects/chess/tests/resources/starting_board.txt
