+++ projects/chess/tests/resources/starting_board_coords.txt
