+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "dh_workspace"
 version = "0.0.1"
-description = "Placeholder Python package converted from a ROS workspace"
+description = "Multi-project prototyping workspace with a chess module"
 readme = "README.md"
 requires-python = ">=3.8"
 keywords = ["example"]
