+++ src/projects/__init__.py
@@ -1,0 +1,3 @@
+"""Namespace package for prototype projects."""
+
+__all__ = ["chess"]
